Var REMOTE_IMAGE and BUILD_PARAMS for docker build
This will allow the main script to remain generic and work in default
case or locally, but allow for customization when invoked if special
cases are needed (as in our current JenkinsX scenario).

This also contains a small fix that has a default value (of 0) for
PUSH_CACHE so that it can be omitted when invoking the script without
a crasher due to POSIX expansion weirdness.