fix(ci): circle fallback to *_TAG when no *_BRANCH
When a workflow was triggered by a tag push, CircleCI does not populate
the $CIRCLE_BRANCH environment variable (or rather, they do populate it,
but with an empty string), and instead populates a $CIRCLE_TAG variable
with the tag name. Since our ci/build.sh script runs both during commit
and tag pushes, we need to have a fallback for what gets passed to the
script invocation.

This should address the issue where during tag push invoked runs (e.g.
upon release), the script was failing with `line 29: git: not found`,
which was caused by the script not getting a valid BRANCH passed and
falling back to it's userland behavior of trying to invoke git.

Test plan:
We should see that this didn't break normal commit pushes as soon as
this commit is pushed to the PR branch, and I'm fairly confident in it,
but we won't know if the fix works 100% until we do another tag release
with this file active.