fix(ci): extend CIRCLE_BRANCH/CIRCLE_TAG toggle
Looks like we need this one other place for CircleCI. Unfortunately,
I'm not 100% certain their YAML parser will allow variable transforms
in this fashion, and I can't think of a way to test short of another
test release after this is merged.

Test Plan:
If this doesnt break normal builds in CircleCI (should see in the PR
checks, verify that is working first), we can merge and try another test
release.