it is now possible to make expressions with domain types.
Right now cross types is not possible (i.e. Amount with Number) but at least within Amount it is. And the alias is being validated!

The last issue I have is adding at least a little bit of context by giving the expression that is problematic. Right now we just say that something is wrong and it can be hard to figure out exactly what