Type plus validation (#186)
* starting to set up type & fix circe error

* fix validation but still working on variable type

* fixed constructor for abstractdomaintype

* compiles

* add some additional cases where we want to handle a defineddomaintype

* a few changes didn't get saved

* making sure we handle the defineddomaintype in various places & overriding some methods for the type

* fix test

* add test to executionengine spec as well

* fix one constructor error

* debugging

* fix validation

* getting all tests passing WIP

* tests getting there

* one additional test passing, one still failing

* committing EOD debugging

* refactor to use Validation not condition & error message

* fix a few errors post merge and experiment with making Validation optional initially"

* testing various approaches WIP

* fixed validation issue in tests

* clean up

* fixed Domain Type
fixed issues in the validation logic
TODO: add some more tests around the new type definition

* missed this DomainInformation -> DomainType

* cleanup test

* it is now possible to make expressions with domain types.

Right now cross types is not possible (i.e. Amount with Number) but at least within Amount it is. And the alias is being validated!

The last issue I have is adding at least a little bit of context by giving the expression that is problematic. Right now we just say that something is wrong and it can be hard to figure out exactly what